+++ lib/capybara_accessible_selectors/selectors/main.rb
@@ -18,6 +18,8 @@ Capybara.add_selector :main, locator_type: [String, Symbol] do
       node[:"aria-label"].public_send(method, locator.to_s)
     end
   end
+
+  filter_set(:capybara_accessible_selectors, %i[described_by])
 end
 
 module CapybaraAccessibleSelectors